+++ libMx/d_ilm_status.c
@@ -10,7 +10,7 @@
  *
  *--------------------------------------------------------------------------
  *
- * Copyright 2007-2008 Illinois Institute of Technology
+ * Copyright 2007-2009 Illinois Institute of Technology
  *
  * See the file "LICENSE" for information on usage and redistribution
  * of this file, and for a DISCLAIMER OF ALL WARRANTIES.
@@ -243,7 +243,7 @@ mxd_ilm_status_open( MX_RECORD *record )
 	if ( mx_status.code != MXE_SUCCESS )
 		return mx_status;
 
-	channel_usage = mx_hex_char_to_unsigned_long(
+	channel_usage = (int) mx_hex_char_to_unsigned_long(
 				response[ilm_status->channel] );
 
 #if MXD_ILM_STATUS_DEBUG
