+++ libMx/mx_math.h
@@ -11,7 +11,7 @@
  *
  *---------------------------------------------------------------------------
  *
- * Copyright 2007 Illinois Institute of Technology
+ * Copyright 2007, 2009 Illinois Institute of Technology
  *
  * See the file "LICENSE" for information on usage and redistribution
  * of this file, and for a DISCLAIMER OF ALL WARRANTIES.
@@ -32,7 +32,7 @@ MX_API void mx_solve_tridiagonal_matrix( double *a,
 					double *c,
 					double *y,
 					double *x,
-					int n );
+					unsigned long n );
 
 typedef struct {
 	unsigned long num_points;
