+++ libMx/mx_relay.h
@@ -7,7 +7,7 @@
  *
  *--------------------------------------------------------------------------
  *
- * Copyright 1999-2002, 2006-2008 Illinois Institute of Technology
+ * Copyright 1999-2002, 2006-2009 Illinois Institute of Technology
  *
  * See the file "LICENSE" for information on usage and redistribution
  * of this file, and for a DISCLAIMER OF ALL WARRANTIES.
@@ -41,7 +41,7 @@ typedef struct {
 
 	long pulse_on_value;
 	long pulse_off_value;
-	long pulse_duration;
+	double pulse_duration;		/* in seconds */
 } MX_RELAY;
 
 #define MXLV_RLY_RELAY_COMMAND	1001
